Encapsulation of the stages of the game/functions
The encapsulation done was to encapsulate the first render function to split it into two render functions each render function now represents each stage of the game which is the start menu, selection menu, and the battle. Then encapsulation was further done by making separate files for each stage and function for easier code readability. Another file was made to encapsulate the pokemon move functions as well to have more organized code and increase code readability. The .h file was also reorganized and modified for encapsulation and organization.

The comments were added to highlight and emphasize important code blocks, definitions, and initialization, along with mechanic insight. Some commented-out blocks of code were reorganized or reformatted so any threat actors cannot abuse the commented code and increase the readability of the rest of the code. Some comments were added to explain how the process of the line of code or function or helpful tips. Removed inappropriate comments. Comments were added to organize and denote sections of the code or to denote the purpose of that function or how it will be seen as.