Deleting a redundant .cpp legacy file
Since the encapsulation has been done. This file has been removed from the cmake debug and any other references to the file have been removed from other files. The file is no longer recognized by the program and may pose a threat to the stability and security of the program. The content of this file is the legacy code that was debuted at the project presentation for prog II.